add feature: setmeal management -> add setmeal

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -122,7 +122,7 @@ public class DishController {
      */
     @GetMapping("/list")
     @ApiOperation("根据分类id查询菜品")
-    public Result<List<Dish>> list(Integer categoryId) {
+    public Result<List<Dish>> list(Long categoryId) {
         log.info("根据分类id查询菜品: {}", categoryId);
         List<Dish> list = dishService.list(categoryId);
         return Result.success(list);
