+++ sky-server/src/main/java/com/sky/service/DishService.java
@@ -66,5 +66,5 @@ public interface DishService {
      * @param categoryId
      * @return
      */
-    List<Dish> list(Integer categoryId);
+    List<Dish> list(Long categoryId);
 }
